Use external patiencediff.
