+++ breezy/diff.py
@@ -25,6 +25,7 @@ import sys
 from .lazy_import import lazy_import
 lazy_import(globals(), """
 import errno
+import patiencediff
 import subprocess
 import tempfile
 
@@ -33,7 +34,6 @@ from breezy import (
     controldir,
     errors,
     osutils,
-    patiencediff,
     textfile,
     timestamp,
     views,
@@ -93,12 +93,11 @@ def internal_diff(old_label, oldlines, new_label, newlines, to_file,
 
     if sequence_matcher is None:
         sequence_matcher = patiencediff.PatienceSequenceMatcher
-    ud = patiencediff.unified_diff_bytes(oldlines, newlines,
-                                         fromfile=old_label.encode(
-                                             path_encoding, 'replace'),
-                                         tofile=new_label.encode(
-                                             path_encoding, 'replace'),
-                                         n=context_lines, sequencematcher=sequence_matcher)
+    ud = unified_diff_bytes(
+            oldlines, newlines,
+            fromfile=old_label.encode( path_encoding, 'replace'),
+            tofile=new_label.encode(path_encoding, 'replace'),
+            n=context_lines, sequencematcher=sequence_matcher)
 
     ud = list(ud)
     if len(ud) == 0:  # Identical contents, nothing to do
@@ -115,6 +114,75 @@ def internal_diff(old_label, oldlines, new_label, newlines, to_file,
         if not line.endswith(b'\n'):
             to_file.write(b"\n\\ No newline at end of file\n")
     to_file.write(b'\n')
+
+
+def unified_diff_bytes(a, b, fromfile=b'', tofile=b'', fromfiledate=b'',
+                       tofiledate=b'', n=3, lineterm=b'\n', sequencematcher=None):
+    r"""
+    Compare two sequences of lines; generate the delta as a unified diff.
+
+    Unified diffs are a compact way of showing line changes and a few
+    lines of context.  The number of context lines is set by 'n' which
+    defaults to three.
+
+    By default, the diff control lines (those with ---, +++, or @@) are
+    created with a trailing newline.  This is helpful so that inputs
+    created from file.readlines() result in diffs that are suitable for
+    file.writelines() since both the inputs and outputs have trailing
+    newlines.
+
+    For inputs that do not have trailing newlines, set the lineterm
+    argument to "" so that the output will be uniformly newline free.
+
+    The unidiff format normally has a header for filenames and modification
+    times.  Any or all of these may be specified using strings for
+    'fromfile', 'tofile', 'fromfiledate', and 'tofiledate'.  The modification
+    times are normally expressed in the format returned by time.ctime().
+
+    Example:
+
+    >>> for line in bytes_unified_diff(b'one two three four'.split(),
+    ...             b'zero one tree four'.split(), b'Original', b'Current',
+    ...             b'Sat Jan 26 23:30:50 1991', b'Fri Jun 06 10:20:52 2003',
+    ...             lineterm=b''):
+    ...     print line
+    --- Original Sat Jan 26 23:30:50 1991
+    +++ Current Fri Jun 06 10:20:52 2003
+    @@ -1,4 +1,4 @@
+    +zero
+     one
+    -two
+    -three
+    +tree
+     four
+    """
+    if sequencematcher is None:
+        sequencematcher = difflib.SequenceMatcher
+
+    if fromfiledate:
+        fromfiledate = b'\t' + bytes(fromfiledate)
+    if tofiledate:
+        tofiledate = b'\t' + bytes(tofiledate)
+
+    started = False
+    for group in sequencematcher(None, a, b).get_grouped_opcodes(n):
+        if not started:
+            yield b'--- %s%s%s' % (fromfile, fromfiledate, lineterm)
+            yield b'+++ %s%s%s' % (tofile, tofiledate, lineterm)
+            started = True
+        i1, i2, j1, j2 = group[0][1], group[-1][2], group[0][3], group[-1][4]
+        yield b"@@ -%d,%d +%d,%d @@%s" % (i1 + 1, i2 - i1, j1 + 1, j2 - j1, lineterm)
+        for tag, i1, i2, j1, j2 in group:
+            if tag == 'equal':
+                for line in a[i1:i2]:
+                    yield b' ' + line
+                continue
+            if tag == 'replace' or tag == 'delete':
+                for line in a[i1:i2]:
+                    yield b'-' + line
+            if tag == 'replace' or tag == 'insert':
+                for line in b[j1:j2]:
+                    yield b'+' + line
 
 
 def _spawn_external_diff(diffcmd, capture_errors=True):
