+++ breezy/multiparent.py
@@ -24,11 +24,11 @@ from .lazy_import import lazy_import
 lazy_import(globals(), """
 import gzip
 import itertools
+import patiencediff
 
 from breezy import (
     bencode,
     errors,
-    patiencediff,
     ui,
     )
 """)
