+++ breezy/tests/features.py
@@ -404,8 +404,6 @@ subunit = ModuleAvailableFeature('subunit')
 testtools = ModuleAvailableFeature('testtools')
 flake8 = ModuleAvailableFeature('flake8.api.legacy')
 
-compiled_patiencediff_feature = ModuleAvailableFeature(
-    'breezy._patiencediff_c')
 lsprof_feature = ModuleAvailableFeature('breezy.lsprof')
 pkg_resources_feature = ModuleAvailableFeature('pkg_resources')
