+++ breezy/tests/test_import_tariff.py
@@ -176,7 +176,6 @@ class TestImportTariffs(ImportTariffTestCase):
             'breezy.merge3',
             'breezy.merge_directive',
             'breezy.msgeditor',
-            'breezy.patiencediff',
             'breezy.bzr.remote',
             'breezy.rules',
             'breezy.sign_my_commits',
@@ -252,7 +251,6 @@ class TestImportTariffs(ImportTariffTestCase):
                                       'breezy.merge3',
                                       'breezy.merge_directive',
                                       'breezy.msgeditor',
-                                      'breezy.patiencediff',
                                       'breezy.bzr.remote',
                                       'breezy.rules',
                                       'breezy.sign_my_commits',
