+++ breezy/tests/test_multiparent.py
@@ -14,11 +14,11 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
+import patiencediff
 from unittest import TestCase
 
 from .. import (
     multiparent,
-    patiencediff,
     tests,
     )
 from ..sixish import int2byte
