+++ breezy/textmerge.py
@@ -21,7 +21,7 @@ from __future__ import absolute_import
 
 from .lazy_import import lazy_import
 lazy_import(globals(), """
-from breezy import patiencediff
+import patiencediff
 """)
